test: rename test for consistency
